move all cypress tests to cypress folder
