+++ cypress/tsconfig.json
@@ -1,10 +1,10 @@
 {
-  "extends": "../../tsconfig.json",
+  "extends": "../tsconfig.json",
   "compilerOptions": {
     "noEmit": true,
     "types": ["cypress", "@testing-library/cypress"],
     "isolatedModules": false
   },
-  "includes": ["./tests/**.spec.ts"],
+  "includes": ["./e2e/**/*.spec.ts", "./component/**/*.tsx"],
   "exclude": []
 }
